refactor(popup): split popup.ts into dom/settings-form/actions modules
popup.ts was a 718-line module mixing element binding, the entire settings
view (persistence, sections MRU, field picker, template previews, autocomplete
wiring) and the four export flows. Adding a setting or an export target meant
editing one giant file and hunting for the relevant block.

Split by responsibility:
  - popup/dom.ts          — shared element references (resolved once)
  - popup/settings-form.ts — settings view: restore, persist, previews, and
    the field-map state; exposes currentFrontmatterFields() for extraction
  - popup/actions.ts       — Download / Copy / PDF / Obsidian flows + the
    status/loading helpers
  - popup.ts               — 42-line entry: footer, i18n, view switch, init*

Behavior is unchanged — code moved verbatim, the only edit being that actions
reads the frontmatter maps through currentFrontmatterFields() instead of
closing over them directly. Now "add a setting" touches settings-form.ts and
"add an export target" touches actions.ts, in isolation.

tsc clean (incl. noUnusedLocals), 111/111 tests pass, popup bundle flat.
